pyproject.toml
[tool.poetry]
name = "manim-website-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^3.0"
python-dotenv = "^0.21"
flask-cors = "^4.0"
certifi = "*"
sqlalchemy = "^1.4.51"
Flask-SQLAlchemy = "^3.0.5"
cryptography = "^42.0.0" # needed by PyMySQL
psycopg2 = "^2.9.9"

[tool.poetry.group.updator.dependencies]
requests = "^2"
feedparser = "^6.0"
google-api-python-client = "^2.51"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
